+++ source_gen/source_gen/lib/src/generator.dart
@@ -7,6 +7,12 @@ import 'package:analyzer/src/generated/element.dart';
 abstract class Generator {
   const Generator();
 
+  /// The files to track when doing incremental generation.
+  ///
+  /// By default, incremental generation is only done on a library if a source
+  /// file in that library is changes. [AssociatedFiles.sameLibrary]
+  AssociatedFileSet get associatedFileSet => AssociatedFileSet.sameLibrary;
+
   Future<String> generate(Element element) => null;
 }
 
@@ -19,4 +25,12 @@ class InvalidGenerationSourceError {
 
   @override
   String toString() => message;
+}
+
+enum AssociatedFileSet {
+  /// Only run incremental generation when a source file in the library changes.
+  sameLibrary,
+  /// Run incremental generation when a source file in the library changes or
+  /// if a file in the same directory as a library source file changes.
+  sameDirectory
 }